+++ src/data-workspace/inputs/index.js
@@ -2,3 +2,4 @@ export * from './file-inputs.js'
 export * from './long-text.js'
 export * from './inputs.js'
 export * from './true-only-checkbox.js'
+export * from './boolean-radios.js'
